search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.13",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]
